Make group scan test cases templated on type combinations
This commit follows the recent templating style for test combinations
introduced on accessors in
https://github.com/KhronosGroup/SYCL-CTS/pull/753 to also reduce the
overhead for section-lookup in group_scan tests.

Additionally this unifies the two type_coverage.h files to reduce the
maintenance burden of the replicated implementations in them.
