+++ tests/accessor/accessor_common.h
@@ -243,129 +243,6 @@ using targets_pack =
     value_pack<sycl::target, sycl::target::device, sycl::target::host_task>;
 
 /**
- * @brief Lightweight struct for containing tuple types with the singleton packs
- *        comprising a single combination to test.
- */
-template <typename... Ts>
-struct combinations_list {};
-
-/**
- * @brief Helper trait for concatenating two combinations_list.
- */
-template <typename LCombs, typename RCombs>
-struct concat_combination_lists;
-template <typename... LCombTs, typename... RCombTs>
-struct concat_combination_lists<combinations_list<LCombTs...>,
-                                combinations_list<RCombTs...>> {
-  using type = combinations_list<LCombTs..., RCombTs...>;
-};
-
-/**
- * @brief Helper trait for appending an element to a combinations_list.
- */
-template <typename Comb, typename Combs>
-struct append_comb;
-template <typename Comb, typename... CombTs>
-struct append_comb<Comb, combinations_list<CombTs...>> {
-  using type = combinations_list<Comb, CombTs...>;
-};
-
-/**
- * @brief Helper trait for appending a type to each element in a
- * combinations_list.
- */
-template <typename T, typename Combs>
-struct append_type_to_combs;
-template <typename T, typename... ElemTs, typename... CombTs>
-struct append_type_to_combs<
-    T, combinations_list<std::tuple<ElemTs...>, CombTs...>> {
-  using tail =
-      typename append_type_to_combs<T, combinations_list<CombTs...>>::type;
-  using type = typename append_comb<std::tuple<ElemTs..., T>, tail>::type;
-};
-template <typename T>
-struct append_type_to_combs<T, combinations_list<>> {
-  using type = combinations_list<>;
-};
-
-/**
- * @brief Helper trait for either appending a type to each element of a
- * combinations_list or creating a new combination with T if the given
- * combination_list is empty.
- */
-template <typename T, typename Combs>
-struct append_or_create_combs {
-  using type = typename append_type_to_combs<T, Combs>::type;
-};
-template <typename T>
-struct append_or_create_combs<T, combinations_list<>> {
-  using type = combinations_list<std::tuple<T>>;
-};
-
-/**
- * @brief Helper trait for getting all combinations of the values and types in
- * the specified packs.
- */
-template <typename CurrentLevelCombs, typename LastLevelCombs,
-          typename... Packs>
-struct get_combinations_helper;
-template <typename CurrentLevelCombs, typename LastLevelCombs>
-struct get_combinations_helper<CurrentLevelCombs, LastLevelCombs> {
-  using type = LastLevelCombs;
-};
-template <typename CurrentLevelCombs, typename LastLevelCombs, typename T, T V,
-          T... Vs, typename... Packs>
-struct get_combinations_helper<CurrentLevelCombs, LastLevelCombs,
-                               value_pack<T, V, Vs...>, Packs...> {
-  using new_combs =
-      typename append_or_create_combs<value_pack<T, V>, LastLevelCombs>::type;
-  using new_current_level_combs =
-      typename concat_combination_lists<CurrentLevelCombs, new_combs>::type;
-  using type =
-      typename get_combinations_helper<new_current_level_combs, LastLevelCombs,
-                                       value_pack<T, Vs...>, Packs...>::type;
-};
-template <typename CurrentLevelCombs, typename LastLevelCombs, typename T,
-          typename... Packs>
-struct get_combinations_helper<CurrentLevelCombs, LastLevelCombs, value_pack<T>,
-                               Packs...> {
-  using type =
-      typename get_combinations_helper<combinations_list<>, CurrentLevelCombs,
-                                       Packs...>::type;
-};
-template <typename CurrentLevelCombs, typename LastLevelCombs, typename T,
-          typename... Ts, typename... Packs>
-struct get_combinations_helper<CurrentLevelCombs, LastLevelCombs,
-                               type_pack<T, Ts...>, Packs...> {
-  using new_combs =
-      typename append_or_create_combs<type_pack<T>, LastLevelCombs>::type;
-  using new_current_level_combs =
-      typename concat_combination_lists<CurrentLevelCombs, new_combs>::type;
-  using type =
-      typename get_combinations_helper<new_current_level_combs, LastLevelCombs,
-                                       type_pack<Ts...>, Packs...>::type;
-};
-template <typename CurrentLevelCombs, typename LastLevelCombs,
-          typename... Packs>
-struct get_combinations_helper<CurrentLevelCombs, LastLevelCombs, type_pack<>,
-                               Packs...> {
-  using type =
-      typename get_combinations_helper<combinations_list<>, CurrentLevelCombs,
-                                       Packs...>::type;
-};
-
-/**
- * @brief Trait for getting all combinations of the values and types in the
- * specified packs.
- */
-template <typename... Packs>
-struct get_combinations {
-  using type =
-      typename get_combinations_helper<combinations_list<>, combinations_list<>,
-                                       Packs...>::type;
-};
-
-/**
  * @brief Function helps to generate type_pack with sycl::vec of all supported
  * sizes
  */
