+++ tests/group_functions/group_functions_common.h
@@ -21,7 +21,7 @@
 #include "../common/common.h"
 #include "../common/get_group_range.h"
 #include "../common/once_per_unit.h"
-#include "type_coverage.h"
+#include "../common/type_coverage.h"
 
 #include <catch2/catch_template_test_macros.hpp>
