+++ tests/group_functions/group_scan_fp16.cpp
@@ -20,7 +20,7 @@
 
 #include "../common/common.h"
 #include "../common/disabled_for_test_case.h"
-#include "type_coverage.h"
+#include "../common/type_coverage.h"
 #if !SYCL_CTS_COMPILING_WITH_HIPSYCL
 #include "group_scan.h"
 
@@ -42,12 +42,26 @@ using ScanTypes = Types;
 using HalfType = unnamed_type_pack<sycl::half>;
 using HalfExtendedTypes = concatenation<ScanTypes, sycl::half>::type;
 
-static const auto Dims = integer_pack<1, 2, 3>::generate_unnamed();
+using Dims = integer_pack<1, 2, 3>;
+
+using TestCombinations1Type = typename get_combinations<Dims, HalfType>::type;
+
+// FIXME: hipSYCL cannot handle cases of different types
+#if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
+using TestCombinations2Types = TestCombinations1Type;
+using TestCombinations3Types = TestCombinations1Type;
+#else
+using TestCombinations2Types =
+    typename get_combinations<Dims, HalfExtendedTypes, HalfExtendedTypes>::type;
+using TestCombinations3Types =
+    typename get_combinations<Dims, HalfExtendedTypes, HalfExtendedTypes,
+                              HalfExtendedTypes>::type;
+#endif
 #endif  // !SYCL_CTS_COMPILING_WITH_HIPSYCL
 // FIXME: known_identity is not impemented yet for hipSYCL.
-DISABLED_FOR_TEST_CASE(hipSYCL)
+DISABLED_FOR_TEMPLATE_LIST_TEST_CASE(hipSYCL)
 ("Group and sub-group joint scan functions",
- "[group_func][type_list][fp16][dim]")({
+ "[group_func][type_list][fp16][dim]", TestCombinations2Types)({
 #if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
   WARN(
       "hipSYCL cannot handle cases of different types for InPtr and OutPtr. "
@@ -66,14 +80,24 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
   return;
 #else
   auto queue = once_per_unit::get_queue();
+
+  // Get the packs from the test combination type.
+  using DimensionsPack = std::tuple_element_t<0, TestType>;
+  using Type1Pack = std::tuple_element_t<1, TestType>;
+  const auto Dimensions = DimensionsPack::generate_unnamed();
+  const auto Types1 = Type1Pack{};
+
   if (queue.get_device().has(sycl::aspect::fp16)) {
     // FIXME: hipSYCL cannot handle cases of different types
 #if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
-    for_all_combinations<invoke_joint_scan_group_same_type>(Dims, HalfType{},
+    for_all_combinations<invoke_joint_scan_group_same_type>(Dimensions, Types1,
                                                             queue);
 #else
+    using Type2Pack = std::tuple_element_t<2, TestType>;
+    const auto Types2 = Type2Pack{};
+
     for_all_combinations_with<invoke_joint_scan_group, sycl::half>(
-        Dims, HalfExtendedTypes{}, HalfExtendedTypes{}, queue);
+        Dimensions, Types1, Types2, queue);
 #endif
   } else {
     WARN("Device does not support half precision floating point operations.");
@@ -82,9 +106,9 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
 });
 
 // FIXME: known_identity is not impemented yet for hipSYCL.
-DISABLED_FOR_TEST_CASE(hipSYCL)
+DISABLED_FOR_TEMPLATE_LIST_TEST_CASE(hipSYCL)
 ("Group and sub-group joint scan functions with init",
- "[group_func][type_list][fp16][dim]")({
+ "[group_func][type_list][fp16][dim]", TestCombinations3Types)({
 #if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
   WARN(
       "hipSYCL cannot handle cases of different types for T, *InPtr and "
@@ -103,15 +127,26 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
   return;
 #else
   auto queue = once_per_unit::get_queue();
+
+  // Get the packs from the test combination type.
+  using DimensionsPack = std::tuple_element_t<0, TestType>;
+  using Type1Pack = std::tuple_element_t<1, TestType>;
+  const auto Dimensions = DimensionsPack::generate_unnamed();
+  const auto Types1 = Type1Pack{};
+
   if (queue.get_device().has(sycl::aspect::fp16)) {
     // FIXME: hipSYCL cannot handle cases of different types
 #if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
-    for_all_combinations<invoke_init_joint_scan_group_same_type>(
-        Dims, HalfType{}, queue);
+    for_all_combinations<invoke_init_joint_scan_group_same_type>(Dimensions,
+                                                                 Types1, queue);
 #else
+    using Type2Pack = std::tuple_element_t<2, TestType>;
+    using Type3Pack = std::tuple_element_t<3, TestType>;
+    const auto Types2 = Type2Pack{};
+    const auto Types3 = Type3Pack{};
+
     for_all_combinations_with<invoke_init_joint_scan_group, sycl::half>(
-        Dims, HalfExtendedTypes{}, HalfExtendedTypes{}, HalfExtendedTypes{},
-        queue);
+        Dimensions, Types1, Types2, Types3, queue);
 #endif
   } else {
     WARN("Device does not support half precision floating point operations.");
@@ -120,15 +155,23 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
 });
 
 // FIXME: known_identity is not impemented yet for hipSYCL.
-DISABLED_FOR_TEST_CASE(hipSYCL)
-("Group and sub-group scan functions", "[group_func][fp16][dim]")({
+DISABLED_FOR_TEMPLATE_LIST_TEST_CASE(hipSYCL)
+("Group and sub-group scan functions", "[group_func][fp16][dim]",
+ TestCombinations1Type)({
 #ifdef SYCL_CTS_COMPILING_WITH_COMPUTECPP
   WARN("ComputeCpp cannot handle half type. Skipping the test.");
   return;
 #else
   auto queue = once_per_unit::get_queue();
+
+  // Get the packs from the test combination type.
+  using DimensionsPack = std::tuple_element_t<0, TestType>;
+  using Type1Pack = std::tuple_element_t<1, TestType>;
+  const auto Dimensions = DimensionsPack::generate_unnamed();
+  const auto Types1 = Type1Pack{};
+
   if (queue.get_device().has(sycl::aspect::fp16)) {
-    for_all_combinations<invoke_scan_over_group>(Dims, HalfType{}, queue);
+    for_all_combinations<invoke_scan_over_group>(Dimensions, Types1, queue);
   } else {
     WARN("Device does not support half precision floating point operations.");
   }
@@ -137,9 +180,9 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
 
 // FIXME: hipSYCL has wrong arguments order for inclusive_scan_over_group: init
 // and op are interchanged. known_identity is not impemented yet.
-DISABLED_FOR_TEST_CASE(hipSYCL)
+DISABLED_FOR_TEMPLATE_LIST_TEST_CASE(hipSYCL)
 ("Group and sub-group scan functions with init",
- "[group_func][type_list][fp16][dim]")({
+ "[group_func][type_list][fp16][dim]", TestCombinations2Types)({
 #if defined(SYCL_CTS_COMPILING_WITH_COMPUTECPP)
   WARN(
       "ComputeCpp does not implement scan for unsigned long long int and "
@@ -155,14 +198,24 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
   return;
 #else
   auto queue = once_per_unit::get_queue();
+
+  // Get the packs from the test combination type.
+  using DimensionsPack = std::tuple_element_t<0, TestType>;
+  using Type1Pack = std::tuple_element_t<1, TestType>;
+  const auto Dimensions = DimensionsPack::generate_unnamed();
+  const auto Types1 = Type1Pack{};
+
   if (queue.get_device().has(sycl::aspect::fp16)) {
     // FIXME: ComputeCpp cannot handle cases of different types
 #if defined(SYCL_CTS_COMPILING_WITH_COMPUTECPP)
-    for_all_combinations<invoke_init_scan_over_group_same_type>(
-        Dims, HalfType{}, queue);
+    for_all_combinations<invoke_init_scan_over_group_same_type>(Dimensions,
+                                                                Types1, queue);
 #else
+    using Type2Pack = std::tuple_element_t<2, TestType>;
+    const auto Types2 = Type2Pack{};
+
     for_all_combinations_with<invoke_init_scan_over_group, sycl::half>(
-        Dims, HalfExtendedTypes{}, HalfExtendedTypes{}, queue);
+        Dimensions, Types1, Types2, queue);
 #endif
   } else {
     WARN("Device does not support half precision floating point operations.");
