+++ tests/group_functions/group_scan_fp64_fp16.cpp
@@ -23,7 +23,7 @@
 
 #include "../common/common.h"
 #include "../common/disabled_for_test_case.h"
-#include "type_coverage.h"
+#include "../common/type_coverage.h"
 #if !SYCL_CTS_COMPILING_WITH_HIPSYCL
 #include "group_scan.h"
 
@@ -45,11 +45,25 @@ using ScanTypes = Types;
 using DoubleHalfTypes = unnamed_type_pack<double, sycl::half>;
 using DoubleHalfExtendedTypes = concatenation<ScanTypes, DoubleHalfTypes>::type;
 
-static const auto Dims = integer_pack<1, 2, 3>::generate_unnamed();
+using Dims = integer_pack<1, 2, 3>;
+
+using TestCombinations2Types =
+    typename get_combinations<Dims, DoubleHalfTypes, DoubleHalfTypes>::type;
+
+// FIXME: hipSYCL cannot handle cases of different types
+#if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
+using TestCombinations3Types = TestCombinations2Types;
+#else
+using TestCombinations3Types =
+    typename get_combinations<Dims, DoubleHalfExtendedTypes,
+                              DoubleHalfExtendedTypes,
+                              DoubleHalfExtendedTypes>::type;
+#endif
 #endif  // !SYCL_CTS_COMPILING_WITH_HIPSYCL
 // FIXME: known_identity is not impemented yet for hipSYCL.
-DISABLED_FOR_TEST_CASE(hipSYCL)
-("Group and sub-group joint scan functions", "[group_func][fp16][fp64][dim]")({
+DISABLED_FOR_TEMPLATE_LIST_TEST_CASE(hipSYCL)
+("Group and sub-group joint scan functions", "[group_func][fp16][fp64][dim]",
+ TestCombinations2Types)({
 #if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
   WARN(
       "hipSYCL cannot handle cases of different types for InPtr and OutPtr. "
@@ -70,12 +84,21 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
   return;
 #else
   auto queue = once_per_unit::get_queue();
+
+  // Get the packs from the test combination type.
+  using DimensionsPack = std::tuple_element_t<0, TestType>;
+  using Type1Pack = std::tuple_element_t<1, TestType>;
+  using Type2Pack = std::tuple_element_t<2, TestType>;
+  const auto Dimensions = DimensionsPack::generate_unnamed();
+  const auto Types1 = Type1Pack{};
+  const auto Types2 = Type2Pack{};
+
   if (queue.get_device().has(sycl::aspect::fp16) &&
       queue.get_device().has(sycl::aspect::fp64)) {
     // here DoubleHalfTypes can be used as only half+double combinations are
     // untested
     for_all_combinations_with_two<invoke_joint_scan_group, double, sycl::half>(
-        Dims, DoubleHalfTypes{}, DoubleHalfTypes{}, queue);
+        Dimensions, Types1, Types2, queue);
   } else {
     WARN(
         "Device does not support half and double precision floating point "
@@ -85,9 +108,9 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
 });
 
 // FIXME: known_identity is not impemented yet for hipSYCL.
-DISABLED_FOR_TEST_CASE(hipSYCL)
+DISABLED_FOR_TEMPLATE_LIST_TEST_CASE(hipSYCL)
 ("Group and sub-group joint scan functions with init",
- "[group_func][type_list][fp16][fp64][dim]")({
+ "[group_func][type_list][fp16][fp64][dim]", TestCombinations3Types)({
 #if defined(SYCL_CTS_COMPILING_WITH_HIPSYCL)
   WARN(
       "hipSYCL cannot handle cases of different types for T, *InPtr and "
@@ -108,12 +131,22 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
   return;
 #else
   auto queue = once_per_unit::get_queue();
+
+  // Get the packs from the test combination type.
+  using DimensionsPack = std::tuple_element_t<0, TestType>;
+  using Type1Pack = std::tuple_element_t<1, TestType>;
+  using Type2Pack = std::tuple_element_t<2, TestType>;
+  using Type3Pack = std::tuple_element_t<3, TestType>;
+  const auto Dimensions = DimensionsPack::generate_unnamed();
+  const auto Types1 = Type1Pack{};
+  const auto Types2 = Type2Pack{};
+  const auto Types3 = Type3Pack{};
+
   if (queue.get_device().has(sycl::aspect::fp16) &&
       queue.get_device().has(sycl::aspect::fp64)) {
     for_all_combinations_with_two<invoke_init_joint_scan_group, double,
-                                  sycl::half>(Dims, DoubleHalfExtendedTypes{},
-                                              DoubleHalfExtendedTypes{},
-                                              DoubleHalfExtendedTypes{}, queue);
+                                  sycl::half>(Dimensions, Types1, Types2,
+                                              Types3, queue);
   } else {
     WARN(
         "Device does not support half and double precision floating point "
@@ -124,9 +157,9 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
 
 // FIXME: hipSYCL has wrong arguments order for inclusive_scan_over_group: init
 // and op are interchanged. known_identity is not impemented yet.
-DISABLED_FOR_TEST_CASE(hipSYCL)
+DISABLED_FOR_TEMPLATE_LIST_TEST_CASE(hipSYCL)
 ("Group and sub-group scan functions with init",
- "[group_func][fp16][fp64][dim]")({
+ "[group_func][fp16][fp64][dim]", TestCombinations2Types)({
 #if defined(SYCL_CTS_COMPILING_WITH_COMPUTECPP)
   WARN(
       "ComputeCpp cannot handle cases of different types for T and V. Skipping "
@@ -140,13 +173,22 @@ DISABLED_FOR_TEST_CASE(hipSYCL)
   return;
 #else
   auto queue = once_per_unit::get_queue();
+
+  // Get the packs from the test combination type.
+  using DimensionsPack = std::tuple_element_t<0, TestType>;
+  using Type1Pack = std::tuple_element_t<1, TestType>;
+  using Type2Pack = std::tuple_element_t<2, TestType>;
+  const auto Dimensions = DimensionsPack::generate_unnamed();
+  const auto Types1 = Type1Pack{};
+  const auto Types2 = Type2Pack{};
+
   if (queue.get_device().has(sycl::aspect::fp16) &&
       queue.get_device().has(sycl::aspect::fp64)) {
     // here DoubleHalfTypes can be used as only half+double combinations are
     // untested
     for_all_combinations_with_two<invoke_init_scan_over_group, double,
-                                  sycl::half>(Dims, DoubleHalfTypes{},
-                                              DoubleHalfTypes{}, queue);
+                                  sycl::half>(Dimensions, Types1, Types2,
+                                              queue);
   } else {
     WARN(
         "Device does not support half and double precision floating point "
